+++ packages/user/src/model/users/handlers/index.ts
@@ -1,0 +1,6 @@
+import changePassword from "./changePassword";
+import me from "./me";
+import updateMe from "./updateMe";
+import users from "./users";
+
+export default { changePassword, me, updateMe, users };