+++ packages/user/src/model/users/resolver.ts
@@ -15,7 +15,11 @@ const Mutation = {
     },
     context: MercuriusContext
   ) => {
-    const service = new Service(context.config, context.database);
+    const service = new Service(
+      context.config,
+      context.database,
+      context.dbSchema
+    );
 
     try {
       return context.user?.id
@@ -49,7 +53,11 @@ const Mutation = {
   ) => {
     const { data } = arguments_;
 
-    const service = new Service(context.config, context.database);
+    const service = new Service(
+      context.config,
+      context.database,
+      context.dbSchema
+    );
 
     try {
       if (context.user?.id) {
@@ -92,13 +100,17 @@ const Mutation = {
 const Query = {
   me: async (
     parent: unknown,
-    arguments_: unknown,
+    arguments_: Record<string, never>,
     context: MercuriusContext
   ) => {
-    const service = new Service(context.config, context.database);
+    const service = new Service(
+      context.config,
+      context.database,
+      context.dbSchema
+    );
 
     if (context.user?.id) {
-      return service.findById(context.user.id);
+      return await service.findById(context.user.id);
     } else {
       context.app.log.error(
         "Could not able to get user id from mercurius context"
@@ -119,7 +131,11 @@ const Query = {
     arguments_: { id: string },
     context: MercuriusContext
   ) => {
-    const service = new Service(context.config, context.database);
+    const service = new Service(
+      context.config,
+      context.database,
+      context.dbSchema
+    );
 
     return await service.findById(arguments_.id);
   },
@@ -134,7 +150,11 @@ const Query = {
     },
     context: MercuriusContext
   ) => {
-    const service = new Service(context.config, context.database);
+    const service = new Service(
+      context.config,
+      context.database,
+      context.dbSchema
+    );
 
     return await service.list(
       arguments_.limit,